+++ src/createKarmaConfig.js
@@ -4,6 +4,7 @@ import merge from 'webpack-merge'
 
 import createWebpackConfig from './createWebpackConfig'
 import debug from './debug'
+import {UserError} from './errors'
 import {deepToString, typeOf} from './utils'
 
 // The following defaults are combined into a single extglob-style pattern to
@@ -165,7 +166,7 @@ export default function createKarmaConfig(args, buildConfig, pluginConfig, userC
     ]
   }
 
-  let karmaConfig = merge({
+  let karmaConfig = {
     browsers,
     coverageReporter: {
       dir: path.resolve('coverage'),
@@ -208,7 +209,22 @@ export default function createKarmaConfig(args, buildConfig, pluginConfig, userC
       noInfo: true,
       quiet: true,
     },
-  }, userKarma.extra)
+  }
+
+  // Any extra user Karma config is merged into the generated config to give
+  // them even more control.
+  if (userKarma.extra) {
+    karmaConfig = merge(karmaConfig, userKarma.extra)
+  }
+
+  // Finally, give them a chance to do whatever they want with the generated
+  // config.
+  if (typeOf(userKarma.config) === 'function') {
+    karmaConfig = userKarma.config(karmaConfig)
+    if (!karmaConfig) {
+      throw new UserError(`karma.config() in ${userConfig.path} didn't return anything - it must return the Karma config object.`)
+    }
+  }
 
   debug('karma config: %s', deepToString(karmaConfig))
   return karmaConfig